# docs/swagger.yaml
basePath: /api
definitions:
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse:
    description: Структура сообщения об ошибке
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response:
    description: Базовый ответ
    properties:
      status:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.CreatePollRequest:
    description: Данные для создания комнаты Planning Poker
    properties:
      card_links:
        items:
          type: string
        type: array
      invitees:
        items:
          type: string
        type: array
    type: object
...
  github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.VotePollRequest:
    description: Оценка (story points) участника голосования
    properties:
      points:
        type: integer
    type: object
host: clac-clac.ru
info:
  contact: {}
  description: Наше API
  title: NeXuS
paths:
...
  /boards/{board_link}/polls:
    delete:
      description: Администратор завершает голосование и удаляет комнату
      parameters:
      - description: UUID доски
        format: uuid
        in: path
        name: board_link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: poll deleted
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response'
        "400":
          description: invalid board link
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "403":
          description: not poll admin
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: poll not found
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: cannot delete poll
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Завершить покер-комнату
      tags:
      - Polls
    get:
      description: Возвращает текущий активный Planning Poker для доски, включая задачи,
        голоса и приглашённых участников
      parameters:
      - description: UUID доски
        format: uuid
        in: path
        name: board_link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: active poll data
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response'
        "400":
          description: invalid board link
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "403":
          description: access to board denied
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: no active poll for this board
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: cannot get active poll
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Получить активный опрос
      tags:
      - Polls
    post:
      consumes:
      - application/json
      description: Создаёт комнату для оценки задач методом Planning Poker. Требует
        прав Admin/Creator.
      parameters:
      - description: UUID доски
        format: uuid
        in: path
        name: board_link
        required: true
        type: string
      - description: Карточки и приглашённые участники
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.CreatePollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: poll created
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response'
        "400":
          description: invalid board link / invalid request schema
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "403":
          description: action denied (not admin/creator)
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "409":
          description: poll already exists
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: cannot create poll
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Создать покер-комнату
      tags:
      - Polls
    put:
      consumes:
      - application/json
      description: Приглашённый участник ставит оценку текущей активной карточке
      parameters:
      - description: UUID доски
        format: uuid
        in: path
        name: board_link
        required: true
        type: string
      - description: Оценка
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_delivery_http_dto.VotePollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: vote accepted
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response'
        "400":
          description: invalid board link / invalid request schema
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "403":
          description: user not invited
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: poll not found
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: cannot vote
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Проголосовать
      tags:
      - Polls
  /boards/{board_link}/polls/next:
    post:
      description: Администратор переходит к следующей карточке в очереди. Если карточек
        больше нет — комната автоматически удаляется.
      parameters:
      - description: UUID доски
        format: uuid
        in: path
        name: board_link
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: moved to next card or poll finished
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.Response'
        "400":
          description: invalid board link
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "403":
          description: not poll admin
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "404":
          description: poll not found
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
        "500":
          description: cannot advance poll
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_Clac_Clac_facade_internal_api.ErrorResponse'
      security:
      - sessionCookie: []
      summary: Следующая карточка
      tags:
      - Polls
...
swagger: "2.0"